requirements.txt
spotipy>=2.24.0
python-dotenv>=1.0.0
requests>=2.32.0
